lambda/moderate_content/handler/requirements.txt
alt-profanity-check==0.24.0
boto3==1.17.66
botocore==1.20.66
jmespath==0.10.0
joblib==1.0.1
numpy==1.20.2
Pillow==8.2.0
python-dateutil==2.8.1
s3transfer==0.4.2
scikit-learn==0.24.0
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
urllib3==1.26.4
uuid==1.30
